refactor: MorphemePosition 위치 이동
