+++ src/main/java/com/almondia/meca/cardhistory/domain/vo/NlpToken.java
@@ -2,6 +2,8 @@ package com.almondia.meca.cardhistory.domain.vo;
 
 import java.util.Objects;
 
+import com.almondia.meca.cardhistory.infra.morpheme.MorphemePosition;
+
 import lombok.Getter;
 
 @Getter
